+++ CamadaDTO/objCongregacao.cs
@@ -3,7 +3,7 @@ namespace CamadaDTO
 {
 	public class objCongregacao
 	{
-		public byte? IDCongregacao { get; set; }
+		public int? IDCongregacao { get; set; }
 		public string Congregacao { get; set; }
 		public bool Ativo { get; set; }
 	}
